fix(console): put Settings MCP on the API the host serves
Opening Settings, MCP Servers threw `Cannot read properties of undefined
(reading 'length')`. The page read `.servers` off a `{ servers }` wrapper
that `GET .../mcp/servers` has never sent (it answers a bare array), stored
the resulting `undefined` as its server list, and threw on render, with the
fetch that caused it nowhere in the stack.

The whole view targeted that phantom surface — servers keyed by `server_id`,
`/connect` and `/disconnect` routes the host does not mount — so it is ported
rather than guarded. `McpServersSection` moves out of `ConnectionsView` into
`views/connections/`, which is now the console's only MCP surface: Connections
renders it inline, Settings renders it standalone. The phantom methods leave
`OpenCompanyClient` and `lib/mcp.ts` goes with them, so the contradictory pair
cannot be picked up again.

The list routes in `api/mcp.ts` now check the one property they promise — that
the body is a list — so the next wire disagreement is a stated load error at
the fetch instead of a TypeError several renders later. The section tells that
apart from a 404: no MCP routes is a fact about the host, an unreadable answer
is not "this company has no servers".
