+++ frontend/src/api/client.ts
@@ -7,12 +7,6 @@
 //     and calls use `/api/v1/companies/{id}/*`.
 
 import type { ConsoleConfig } from "../config";
-import type {
-  InstallMcpServerInput,
-  McpCallResult,
-  McpServer,
-  McpTool,
-} from "../lib/mcp";
 import {
   ApiError,
   type ApiErrorBody,
@@ -569,67 +563,15 @@ export class OpenCompanyClient {
     );
   }
 
-  /** Installed MCP servers plus their live connection state. */
-  listMcpServers(company?: string | null): Promise<{ servers: McpServer[] }> {
-    return this.request<{ servers: McpServer[] }>("GET", `${this.scope(company)}/mcp/servers`);
-  }
-
-  /** Persist a manual stdio or streamable-HTTP MCP server install. */
-  installMcpServer(
-    input: InstallMcpServerInput,
-    company?: string | null,
-  ): Promise<{ server_id: string }> {
-    return this.request<{ server_id: string }>("POST", `${this.scope(company)}/mcp/servers`, input);
-  }
-
-  /** Connect an installed server and return its advertised tools. */
-  connectMcpServer(serverId: string, company?: string | null): Promise<{ tools: McpTool[] }> {
-    return this.request<{ tools: McpTool[] }>(
-      "POST",
-      `${this.scope(company)}/mcp/servers/${encodeURIComponent(serverId)}/connect`,
-    );
-  }
-
-  /** Stop an installed server's live connection. */
-  disconnectMcpServer(
-    serverId: string,
-    company?: string | null,
-  ): Promise<{ disconnected: boolean }> {
-    return this.request<{ disconnected: boolean }>(
-      "POST",
-      `${this.scope(company)}/mcp/servers/${encodeURIComponent(serverId)}/disconnect`,
-    );
-  }
-
-  /** Disconnect and remove an MCP install. */
-  uninstallMcpServer(serverId: string, company?: string | null): Promise<void> {
-    return this.request<void>(
-      "DELETE",
-      `${this.scope(company)}/mcp/servers/${encodeURIComponent(serverId)}`,
-    );
-  }
-
-  /** Cached tools advertised by a connected MCP server. */
-  listMcpTools(serverId: string, company?: string | null): Promise<{ tools: McpTool[] }> {
-    return this.request<{ tools: McpTool[] }>(
-      "GET",
-      `${this.scope(company)}/mcp/servers/${encodeURIComponent(serverId)}/tools`,
-    );
-  }
-
-  /** Invoke one tool through a connected MCP server. */
-  callMcpTool(
-    serverId: string,
-    tool: string,
-    arguments_: Record<string, unknown>,
-    company?: string | null,
-  ): Promise<McpCallResult> {
-    return this.request<McpCallResult>(
-      "POST",
-      `${this.scope(company)}/mcp/servers/${encodeURIComponent(serverId)}/tools/${encodeURIComponent(tool)}/call`,
-      { arguments: arguments_ },
-    );
-  }
+  // The company's MCP tool servers are NOT reachable from this class. They live
+  // in `api/mcp.ts`, as standalone functions over the shared client, and that is
+  // the only MCP surface the console has. A second set of methods used to sit
+  // here — `listMcpServers` promising a `{ servers }` wrapper, servers keyed by
+  // `server_id`, `/connect` and `/disconnect` routes — none of which any host
+  // has ever served. `request<T>` casts an unparsed body to `T`, so those types
+  // were never checked against the wire and the view built on them crashed on
+  // open (issue #414). Add MCP calls to `api/mcp.ts`, next to the ones the host
+  // answers.
 
   /** Platform lifecycle control (requires a scoped company id). */
   lifecycle(action: LifecycleAction, company?: string | null): Promise<CompanyStatus> {
